tuned parameters, made the visualizer show the manouverable areas, updated obstacles and path planner's map, demo ready
